# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Personal Memory API",
    "description": "Private memory and project API. Prefer runAssistantCommand for GPT use; it handles commands, active project context, tasks, notes, meetings, memories, search, summaries, and delete.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://memory-green-kappa.vercel.app"
    }
  ],
  "security": [
    {
      "apiKeyAuth": []
    }
  ],
  "paths": {
...
    "/api/memories/search": {
      "get": {
        "operationId": "searchMemories",
        "summary": "Search memories",
        "description": "Direct memory search. For grouped search across projects, tasks, notes, meetings, and memories, use runAssistantCommand.",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": true,
            "description": "Important words from the user request, such as 'JWT notes' or 'Shubham Mane interview'.",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Search results.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SearchMemoriesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/memories/{id}": {
      "get": {
        "operationId": "getMemoryById",
        "summary": "Get a memory by ID",
        "description": "Retrieve the full details of a single memory.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the memory.",
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Memory details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      },
      "delete": {
        "operationId": "deleteMemory",
        "summary": "Delete a memory",
        "description": "Delete a memory by ID.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the memory.",
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Memory deleted successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteMemoryResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/projects/{id}": {
      "get": {
        "operationId": "getProjectById",
        "summary": "Get a project by ID",
        "description": "Retrieve a saved project.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the project.",
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project details.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
    "/api/projects/{id}/memories": {
      "get": {
        "operationId": "listProjectMemories",
        "summary": "List project memories",
        "description": "List notes, tasks, work, requirements, and credentials saved under one project.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "MongoDB ObjectId of the project.",
            "schema": {
              "type": "string",
              "pattern": "^[a-fA-F0-9]{24}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Project memories.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectMemoriesResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "Memory": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "_id",
          "title",
          "content",
          "category",
          "kind",
          "tags",
          "source",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "_id": {
            "type": "string",
            "description": "MongoDB ObjectId for the memory."
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "category": {
            "type": "string",
            "default": "general"
          },
          "kind": {
            "type": "string",
            "enum": [
              "note",
              "task",
              "work_done",
              "requirement",
              "credential"
            ],
            "default": "note"
          },
          "projectId": {
            "oneOf": [
              {
                "type": "string",
                "pattern": "^[a-fA-F0-9]{24}$"
              },
              {
                "$ref": "#/components/schemas/ProjectSummary"
              }
            ]
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "source": {
            "type": "string",
            "default": "manual"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "__v": {
            "type": "integer"
          }
        }
      },
...
      "ProjectSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "_id",
          "name"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "Project": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "_id",
          "name",
          "description",
          "status",
          "tags",
          "source",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "active",
              "paused",
              "completed",
              "archived"
            ],
            "default": "active"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "source": {
            "type": "string",
            "default": "manual"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "__v": {
            "type": "integer"
          }
        }
      },
      "MemoryResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Memory"
          }
        }
      },
...
      "ProjectResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Project"
          }
        }
      },
...
      "ProjectMemoriesResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "count",
          "project",
          "data"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "minimum": 0
          },
          "project": {
            "$ref": "#/components/schemas/Project"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Memory"
            }
          }
        }
      },
      "SearchMemoriesResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "query",
          "count",
          "limit",
          "data"
        ],
        "properties": {
          "query": {
            "type": "string"
          },
          "count": {
            "type": "integer",
            "minimum": 0
          },
          "limit": {
            "type": "integer",
            "const": 20
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Memory"
            }
          }
        }
      },
      "DeleteMemoryResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "message",
          "data"
        ],
        "properties": {
          "message": {
            "type": "string",
            "example": "Memory deleted"
          },
          "data": {
            "$ref": "#/components/schemas/Memory"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Missing or invalid API key.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal server error.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    }
  }
}